Add a new "ff" prefix
"ff" stands for "fast-forward". It applies a huge (9999) numeric
argument to the next command.

To select the last element of a huge list, do "ffj".
To select the first element of a huge list, do "ffk".

This is due to a recent change in `worf-up' and `worf-down' that forbids
them to step out of the parent outline (just like in lispy).