worf.el: "3T" allows to clock an amount of minutes before now
Let's say I was interrupted by a TASK and could not clock in.

It took me X minutes to finish the TASK and right after I have access to
the clock once more.

So I "3T" and enter X as the amount of minutes.  This way, it's as if I
clocked it right at the moment when I was interrupted.