worf.el (worf-log-time): New defun
It's now possible to "complete a task in the past" with "2T" without
breaking the current clock, i.e. `org-clock-in' and `org-clock-out' are
not used any more.

Also, the `CLOSED' timestamp will be in the past, not the current time.

For example, starting with:

    * TODO Task
    DEADLINE: <2017-04-22 Sat 09:15>

Press "2T", enter 150 to denote the amount of minutes spent on the
task. Result:

    * DONE Task
    CLOSED: [2017-04-22 Sat 11:45] DEADLINE: <2017-04-22 Sat 09:15>
    :LOGBOOK:
    CLOCK: [2017-04-22 Sat 09:15]--[2017-04-22 Sat 11:45] =>  2:30
    :END:

Note that the current time doesn't matter at all: only the deadline time
and the time difference are used.